Bunch of small changes, still works though

+++ CathLab/Inventory/tTypes.aspx.cs
@@ -11,6 +11,8 @@ namespace CathLab
 {
     public partial class Types : System.Web.UI.Page
     {
+        public static string pnum;
+
         protected void Page_Load(object sender, EventArgs e)
         {
             if (!Page.IsPostBack)
@@ -39,22 +41,42 @@ namespace CathLab
 
         protected void rgManufacturers_UpdateCommand(object sender, Telerik.Web.UI.GridCommandEventArgs e)
         {
-            GridEditableItem editedItem = e.Item as GridEditableItem;
-            int a = editedItem.Controls.Count;
+            using (var context = new cathlabEntities())
+            {
+                int pnum = int.Parse(e.Item.OwnerTableView.DataKeyValues[e.Item.ItemIndex]["ID"].ToString());
+                ProductType pt = context.ProductTypes.Find(ID);
+            }
+            
         }
 
         protected void rgPartNumbers_UpdateCommand(object sender, GridCommandEventArgs e)
         {
-            GridEditableItem editedItem = e.Item as GridEditableItem;
+            if (e.CommandName == "Update")
+            {
+                using (var context = new cathlabEntities())
+                {
+                    int PNum = int.Parse(e.Item.OwnerTableView.DataKeyValues[e.Item.ItemIndex]["PartNum"].ToString());
+                    PartNumber pn = context.PartNumbers.Find(PNum);
+                    pn.NameSize = (e.Item.FindControl("tbNameSize") as RadTextBox).Text;
+                    pn.Cost = int.Parse((e.Item.FindControl("tbCost") as RadTextBox).Text);
+                    pn.Par = int.Parse((e.Item.FindControl("tbPar") as RadTextBox).Text);
+                    context.SaveChanges();
+                }
+            }
         }
 
         protected void rgProdType_UpdateCommand(object sender, GridCommandEventArgs e)
         {
-            GridEditableItem editedItem = e.Item as GridEditableItem;
+            using (var context = new cathlabEntities())
+            {
+                int ID = int.Parse(e.Item.OwnerTableView.DataKeyValues[e.Item.ItemIndex]["ID"].ToString());
+                ProductType pt = context.ProductTypes.Find(ID);
+                pt.Type = (e.Item.FindControl("tbType") as RadTextBox).Text;
+                context.SaveChanges();
+            }
         }
 
         #region Delete
-
         protected void rgPartNumbers_DeleteCommand(object sender, GridCommandEventArgs e)
         {
             GridDataItem item = (GridDataItem)e.Item;
@@ -102,53 +124,67 @@ namespace CathLab
             using (var context = new cathlabEntities())
             {
                 ProductType pt = context.ProductTypes.Find(ID);
-                try
+                var temp = (from prod in context.Products
+                           where prod.PartNumber1.ProductTypeID == pt.ID
+                           select prod).FirstOrDefault();
+
+                if(temp ==null)
                 {
                     context.ProductTypes.Remove(pt);
                     context.SaveChanges();
+                    rgProdType.Rebind();
                 }
-                catch (DbUpdateException)
+                else
                 {
                     rnLabel.Text = "ERROR! Can't delete product type. Products in database rely on this entry.";
                     RadNotification.Show();
                 }
+
+                //try
+                //{
+                //    context.ProductTypes.Remove(pt);
+                //    context.SaveChanges();
+                //}
+                //catch (Exception)
+                //{
+                //    rnLabel.Text = "ERROR! Can't delete product type. Products in database rely on this entry.";
+                //    RadNotification.Show();
+                //}
             }
         }
-
-        #endregion Delete
-
-        #region Visibility
-        protected void btnPartNums_Click(object sender, EventArgs e)
-        {
-            rgPartNumbers.Visible = true;
-            rgProdType.Visible = false;
-            rgManufacturers.Visible = false;
-            rgPartNumbers.Rebind();
-        }
-
-        protected void btnProdTypes_Click(object sender, EventArgs e)
-        {
-            rgProdType.Visible = true;
-            rgManufacturers.Visible = false;
-            rgPartNumbers.Visible = false;
-            rgProdType.Rebind();
-        }
-
-        protected void btnManufacturers_Click(object sender, EventArgs e)
-        {
-            rgManufacturers.Visible = true;
-            rgPartNumbers.Visible = false;
-            rgProdType.Visible = false;
-            rgManufacturers.Rebind();
-        }
-
-        #endregion Visibility
+        #endregion Delete       
 
         protected void tbPNSearch_TextChanged(object sender, EventArgs e)
         {
-            using (var context = new cathlabEntities())          
-                rgPartNumbers.DataSource = (from pnum in context.PartNumbers select new { pnum.PartNum, pnum.NameSize, pnum.ProductType.Type, pnum.Cost, pnum.Manufacturer.Name, pnum.Par })
-                    .Where(a => a.PartNum.Contains(tbPNSearch.Text)).ToList();            
+            if (tbPNSearch.Text != string.Empty)
+            {
+                using (var context = new cathlabEntities())
+                    rgPartNumbers.DataSource = (from pnum in context.PartNumbers select new { pnum.PartNum, pnum.NameSize, pnum.ProductType.Type, pnum.Cost, pnum.Manufacturer.Name, pnum.Par })
+                        .Where(a => a.PartNum.Contains(tbPNSearch.Text)).ToList();
+            }
+            else
+                rgPartNumbers.Rebind();
+        }
+
+        protected void btnNewPartNum_Click(object sender, EventArgs e)
+        {
+            //pnlNewPartNumber.Visible = true;
+            //pnlNewManufacturer.Visible = false;
+            //pnlNewProdType.Visible = false;
+        }
+
+        protected void btnManufacturer_Click(object sender, EventArgs e)
+        {
+            //pnlNewPartNumber.Visible = false;
+            //pnlNewManufacturer.Visible = true;
+            //pnlNewProdType.Visible = false;
+        }
+
+        protected void btnProductType_Click(object sender, EventArgs e)
+        {
+            //pnlNewPartNumber.Visible = false;
+            //pnlNewManufacturer.Visible = false;
+            //pnlNewProdType.Visible = true;
         }
     }
 }
